refactor(S-3.02): extract handle_view to assets/view.rs
Moves handle_view (single object lookup + attribute display) to view.rs
shard. All 612 unit tests pass.
